File Renames
- Updated file names to be more accurate:
  - The Notification files were renamed to Hike
  - The Home files were renamed to UserInfo
- Updated the Gradle version to the most recent
- Enabled data binding
