+++ app/build.gradle
@@ -27,6 +27,7 @@ android {
     }
     buildFeatures {
         viewBinding true
+        dataBinding true
     }
 }
