+++ app/src/main/java/com/example/myapplication/ui/hike/HikeViewModel.java
@@ -1,16 +1,16 @@
-package com.example.myapplication.ui.notifications;
+package com.example.myapplication.ui.hike;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class HikeViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public HikeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is hike fragment");
     }
 
     public LiveData<String> getText() {
