+++ app/src/main/java/com/example/myapplication/ui/userInfo/UserInfoViewModel.java
@@ -1,16 +1,16 @@
-package com.example.myapplication.ui.home;
+package com.example.myapplication.ui.userInfo;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class UserInfoViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public UserInfoViewModel() {
         mText = new MutableLiveData<>();
-        //mText.setValue("This is home fragment");
+        //mText.setValue("This is userinfo fragment");
     }
 
     public LiveData<String> getText() {
