+++ app/src/main/res/layout/fragment_userinfo.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.HomeFragment">
+    tools:context=".ui.userInfo.UserInfoFragment">
 
     <de.hdodenhof.circleimageview.CircleImageView
         android:id="@+id/iv_pic"
